Ensure that we don't totally clobber the escape key
Othewise, escape for normal stuff, such as closing the command palette doesn't work. `event.abortKeyBinding()` ensures that the key press will still move on to other handlers.

Note that I did have to remove the activation command, ensuring that this package is activated right away. Otherwise, if it's not activated our code that aborts the keybinding doesn't run yet, and the escape key will still be clobbered.